- moves main writers into own namespaces
