+++ src/Kiota.Builder/Writers/Java/JavaWriter.cs
@@ -1,18 +1,15 @@
-﻿using Kiota.Builder.Writers.CSharp;
-
-namespace Kiota.Builder.Writers
+namespace Kiota.Builder.Writers.Java
 {
-    public class CSharpWriter : LanguageWriter
+    public class JavaWriter : LanguageWriter
     {
-        public CSharpWriter(string rootPath, string clientNamespaceName)
+        public JavaWriter(string rootPath, string clientNamespaceName)
         {
-            PathSegmenter = new CSharpPathSegmenter(rootPath, clientNamespaceName);
-            var conventionService = new CSharpConventionService();
+            PathSegmenter = new JavaPathSegmenter(rootPath, clientNamespaceName);
+            var conventionService = new JavaConventionService();
             Writers = new() {
                 { typeof(CodeClass.Declaration), new CodeClassDeclarationWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
-                { typeof(CodeClass.End), new CodeClassEndWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
+                { typeof(CodeClass.End), new CodeClassEndWriter() as object as ICodeElementWriter<CodeElement> },
                 { typeof(CodeEnum), new CodeEnumWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
-                { typeof(CodeIndexer), new CodeIndexerWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
                 { typeof(CodeMethod), new CodeMethodWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
                 { typeof(CodeProperty), new CodePropertyWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
                 { typeof(CodeType), new CodeTypeWriter(conventionService) as object as ICodeElementWriter<CodeElement> },
